@@ -1,14 +1,14 @@
 {
     "name": "@plexauth/sso",
-    "version": "2.1.3",
+    "version": "2.1.4",
     "private": true,
     "dependencies": {
-        "bootstrap": "^5.1.3",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2"
+        "bootstrap": "^5.2.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0"
     },
     "devDependencies": {
-        "react-scripts": "^5.0.0"
+        "react-scripts": "^5.0.1"
     },
     "scripts": {
         "start": "react-scripts start",
